Partition tests block across config submodules
The tests block in src/config/mod.rs ballooned to ~2,000 lines after the
file split; nearly every test exercises a single submodule's logic, but
they were all forced through one shared scope. Move each test into the
submodule it actually exercises so the file responsible for the type
also owns the assertions.

src/config/mod.rs drops from 2,673 to 965 lines. mod.rs keeps the
14 cross-cutting tests (Config::default, full TOML parse, partial
merging, meeting-mode overrides). The other 75 tests move to:

  hotkey.rs        4 tests
  status.rs       12 tests
  whisper.rs      12 tests
  language.rs      5 tests
  output.rs       20 tests
  profile.rs       6 tests
  meeting.rs       8 tests
  engines/parakeet.rs  5 tests
  engines/mod.rs       3 tests

cargo test --lib still reports 770 passing.
