+++ src/config/hotkey.rs
@@ -83,3 +83,143 @@ pub(super) fn default_hotkey_key() -> String {
         "SCROLLLOCK".to_string()
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::config::Config;
+
+    #[test]
+    fn test_parse_hotkey_disabled_without_key() {
+        // Regression test for GitHub issue #17
+        // When hotkey is disabled, the key field should not be required
+        let toml_str = r#"
+            [hotkey]
+            enabled = false
+
+            [audio]
+            device = "default"
+            sample_rate = 16000
+            max_duration_secs = 60
+
+            [whisper]
+            model = "base.en"
+            language = "en"
+
+            [output]
+            mode = "type"
+        "#;
+
+        let config: Config = toml::from_str(toml_str).unwrap();
+        assert!(!config.hotkey.enabled);
+        assert_eq!(config.hotkey.key, default_hotkey_key()); // platform default
+    }
+
+    #[test]
+    fn test_parse_toggle_mode() {
+        let toml_str = r#"
+            [hotkey]
+            key = "F13"
+            mode = "toggle"
+
+            [audio]
+            device = "default"
+            sample_rate = 16000
+            max_duration_secs = 60
+
+            [audio.feedback]
+            enabled = true
+            theme = "subtle"
+            volume = 0.5
+
+            [whisper]
+            model = "base.en"
+            language = "en"
+
+            [output]
+            mode = "type"
+        "#;
+
+        let config: Config = toml::from_str(toml_str).unwrap();
+        assert_eq!(config.hotkey.key, "F13");
+        assert_eq!(config.hotkey.mode, ActivationMode::Toggle);
+        assert!(config.audio.feedback.enabled);
+        assert_eq!(config.audio.feedback.theme, "subtle");
+        assert_eq!(config.audio.feedback.volume, 0.5);
+    }
+
+    #[test]
+    fn test_parse_profile_modifiers() {
+        let toml_str = r#"
+            [hotkey]
+            key = "SCROLLLOCK"
+
+            [hotkey.profile_modifiers]
+            LEFTSHIFT = "translate"
+            RIGHTALT = "formal"
+
+            [audio]
+            device = "default"
+            sample_rate = 16000
+            max_duration_secs = 60
+
+            [whisper]
+            model = "base.en"
+            language = "en"
+
+            [output]
+            mode = "type"
+
+            [profiles.translate]
+            post_process_command = "translate.sh"
+
+            [profiles.formal]
+            post_process_command = "formal.sh"
+            post_process_timeout_ms = 15000
+        "#;
+
+        let config: Config = toml::from_str(toml_str).unwrap();
+        assert_eq!(config.hotkey.profile_modifiers.len(), 2);
+        assert_eq!(
+            config.hotkey.profile_modifiers.get("LEFTSHIFT").unwrap(),
+            "translate"
+        );
+        assert_eq!(
+            config.hotkey.profile_modifiers.get("RIGHTALT").unwrap(),
+            "formal"
+        );
+        assert!(config.get_profile("translate").is_some());
+        assert!(config.get_profile("formal").is_some());
+        assert_eq!(
+            config
+                .get_profile("translate")
+                .unwrap()
+                .post_process_command
+                .as_deref(),
+            Some("translate.sh")
+        );
+    }
+
+    #[test]
+    fn test_profile_modifiers_default_empty() {
+        let toml_str = r#"
+            [hotkey]
+            key = "SCROLLLOCK"
+
+            [audio]
+            device = "default"
+            sample_rate = 16000
+            max_duration_secs = 60
+
+            [whisper]
+            model = "base.en"
+            language = "en"
+
+            [output]
+            mode = "type"
+        "#;
+
+        let config: Config = toml::from_str(toml_str).unwrap();
+        assert!(config.hotkey.profile_modifiers.is_empty());
+    }
+}
