+++ src/config/mod.rs
@@ -49,8 +49,6 @@ pub use whisper::{WhisperConfig, WhisperMode};
 
 #[cfg(test)]
 use hotkey::default_hotkey_key;
-#[cfg(test)]
-use status::load_icon_theme;
 
 pub(super) fn default_true() -> bool {
     true
@@ -771,1526 +769,6 @@ mod tests {
     }
 
     #[test]
-    fn test_parse_hotkey_disabled_without_key() {
-        // Regression test for GitHub issue #17
-        // When hotkey is disabled, the key field should not be required
-        let toml_str = r#"
-            [hotkey]
-            enabled = false
-
-            [audio]
-            device = "default"
-            sample_rate = 16000
-            max_duration_secs = 60
-
-            [whisper]
-            model = "base.en"
-            language = "en"
-
-            [output]
-            mode = "type"
-        "#;
-
-        let config: Config = toml::from_str(toml_str).unwrap();
-        assert!(!config.hotkey.enabled);
-        assert_eq!(config.hotkey.key, default_hotkey_key()); // platform default
-    }
-
-    #[test]
-    fn test_parse_toggle_mode() {
-        let toml_str = r#"
-            [hotkey]
-            key = "F13"
-            mode = "toggle"
-
-            [audio]
-            device = "default"
-            sample_rate = 16000
-            max_duration_secs = 60
-
-            [audio.feedback]
-            enabled = true
-            theme = "subtle"
-            volume = 0.5
-
-            [whisper]
-            model = "base.en"
-            language = "en"
-
-            [output]
-            mode = "type"
-        "#;
-
-        let config: Config = toml::from_str(toml_str).unwrap();
-        assert_eq!(config.hotkey.key, "F13");
-        assert_eq!(config.hotkey.mode, ActivationMode::Toggle);
-        assert!(config.audio.feedback.enabled);
-        assert_eq!(config.audio.feedback.theme, "subtle");
-        assert_eq!(config.audio.feedback.volume, 0.5);
-    }
-
-    #[test]
-    fn test_parse_auto_submit() {
-        let toml_str = r#"
-            [hotkey]
-            key = "SCROLLLOCK"
-
-            [audio]
-            device = "default"
-            sample_rate = 16000
-            max_duration_secs = 60
-
-            [whisper]
-            model = "base.en"
-            language = "en"
-
-            [output]
-            mode = "type"
-            auto_submit = true
-        "#;
-
-        let config: Config = toml::from_str(toml_str).unwrap();
-        assert!(config.output.auto_submit);
-    }
-
-    #[test]
-    fn test_parse_auto_submit_defaults_false() {
-        let toml_str = r#"
-            [hotkey]
-            key = "SCROLLLOCK"
-
-            [audio]
-            device = "default"
-            sample_rate = 16000
-            max_duration_secs = 60
-
-            [whisper]
-            model = "base.en"
-            language = "en"
-
-            [output]
-            mode = "type"
-        "#;
-
-        let config: Config = toml::from_str(toml_str).unwrap();
-        assert!(!config.output.auto_submit);
-    }
-
-    #[test]
-    fn test_builtin_icon_themes() {
-        // Test all built-in themes load correctly
-        let themes = [
-            "emoji",
-            "nerd-font",
-            "material",
-            "phosphor",
-            "codicons",
-            "omarchy",
-            "minimal",
-            "dots",
-            "arrows",
-            "text",
-        ];
-
-        for theme in themes {
-            let icons = load_icon_theme(theme);
-            assert!(
-                !icons.idle.is_empty() || theme == "emoji",
-                "Theme {} should have idle icon",
-                theme
-            );
-            assert!(
-                !icons.recording.is_empty(),
-                "Theme {} should have recording icon",
-                theme
-            );
-            assert!(
-                !icons.transcribing.is_empty(),
-                "Theme {} should have transcribing icon",
-                theme
-            );
-            // stopped can be empty for some themes
-        }
-    }
-
-    #[test]
-    fn test_emoji_theme_icons() {
-        let icons = load_icon_theme("emoji");
-        assert!(icons.idle.contains("🎙"));
-        assert!(icons.recording.contains("🎤"));
-        assert!(icons.transcribing.contains("⏳"));
-        assert!(icons.stopped.is_empty());
-    }
-
-    #[test]
-    fn test_text_theme_icons() {
-        let icons = load_icon_theme("text");
-        assert_eq!(icons.idle, "[MIC]");
-        assert_eq!(icons.recording, "[REC]");
-        assert_eq!(icons.transcribing, "[...]");
-        assert_eq!(icons.stopped, "[OFF]");
-    }
-
-    #[test]
-    fn test_minimal_theme_icons() {
-        let icons = load_icon_theme("minimal");
-        assert_eq!(icons.idle, "○");
-        assert_eq!(icons.recording, "●");
-        assert_eq!(icons.transcribing, "◐");
-        assert_eq!(icons.stopped, "×");
-    }
-
-    #[test]
-    fn test_status_config_default() {
-        let status = StatusConfig::default();
-        assert_eq!(status.icon_theme, "emoji");
-        assert!(status.icons.idle.is_none());
-        assert!(status.icons.recording.is_none());
-    }
-
-    #[test]
-    fn test_status_config_resolve_icons() {
-        let status = StatusConfig {
-            icon_theme: "text".to_string(),
-            icons: StatusIconOverrides::default(),
-        };
-        let icons = status.resolve_icons();
-        assert_eq!(icons.idle, "[MIC]");
-        assert_eq!(icons.recording, "[REC]");
-    }
-
-    #[test]
-    fn test_status_config_icon_overrides() {
-        let status = StatusConfig {
-            icon_theme: "emoji".to_string(),
-            icons: StatusIconOverrides {
-                idle: None,
-                recording: Some("🔴".to_string()),
-                streaming: None,
-                transcribing: None,
-                stopped: Some("⚫".to_string()),
-            },
-        };
-        let icons = status.resolve_icons();
-        // idle should be from emoji theme
-        assert!(icons.idle.contains("🎙"));
-        // recording should be overridden
-        assert_eq!(icons.recording, "🔴");
-        // transcribing should be from emoji theme
-        assert!(icons.transcribing.contains("⏳"));
-        // stopped should be overridden
-        assert_eq!(icons.stopped, "⚫");
-    }
-
-    #[test]
-    fn test_parse_status_config_from_toml() {
-        let toml_str = r#"
-            [hotkey]
-            key = "SCROLLLOCK"
-
-            [audio]
-            device = "default"
-            sample_rate = 16000
-            max_duration_secs = 60
-
-            [whisper]
-            model = "base.en"
-            language = "en"
-
-            [output]
-            mode = "type"
-
-            [status]
-            icon_theme = "nerd-font"
-        "#;
-
-        let config: Config = toml::from_str(toml_str).unwrap();
-        assert_eq!(config.status.icon_theme, "nerd-font");
-    }
-
-    #[test]
-    fn test_parse_status_icon_overrides_from_toml() {
-        let toml_str = r#"
-            [hotkey]
-            key = "SCROLLLOCK"
-
-            [audio]
-            device = "default"
-            sample_rate = 16000
-            max_duration_secs = 60
-
-            [whisper]
-            model = "base.en"
-            language = "en"
-
-            [output]
-            mode = "type"
-
-            [status]
-            icon_theme = "emoji"
-
-            [status.icons]
-            recording = "🔴"
-            stopped = "⚫"
-        "#;
-
-        let config: Config = toml::from_str(toml_str).unwrap();
-        assert_eq!(config.status.icon_theme, "emoji");
-        assert!(config.status.icons.idle.is_none());
-        assert_eq!(config.status.icons.recording, Some("🔴".to_string()));
-        assert!(config.status.icons.transcribing.is_none());
-        assert_eq!(config.status.icons.stopped, Some("⚫".to_string()));
-
-        let icons = config.status.resolve_icons();
-        assert_eq!(icons.recording, "🔴");
-        assert_eq!(icons.stopped, "⚫");
-    }
-
-    #[test]
-    fn test_invalid_theme_falls_back_to_emoji() {
-        // Non-existent file path should fall back to emoji
-        let icons = load_icon_theme("/nonexistent/path/theme.toml");
-        assert!(icons.idle.contains("🎙"));
-    }
-
-    #[test]
-    fn test_custom_theme_file() {
-        use std::io::Write;
-
-        // Create a temporary theme file
-        let mut temp_file = tempfile::NamedTempFile::new().unwrap();
-        writeln!(
-            temp_file,
-            r#"
-            idle = "IDLE"
-            recording = "REC"
-            transcribing = "BUSY"
-            stopped = "OFF"
-        "#
-        )
-        .unwrap();
-
-        let icons = load_icon_theme(temp_file.path().to_str().unwrap());
-        assert_eq!(icons.idle, "IDLE");
-        assert_eq!(icons.recording, "REC");
-        assert_eq!(icons.transcribing, "BUSY");
-        assert_eq!(icons.stopped, "OFF");
-    }
-
-    #[test]
-    fn test_custom_theme_file_partial() {
-        use std::io::Write;
-
-        // Create a theme file with only some icons (others should default to emoji)
-        let mut temp_file = tempfile::NamedTempFile::new().unwrap();
-        writeln!(
-            temp_file,
-            r#"
-            recording = "🔴"
-        "#
-        )
-        .unwrap();
-
-        let icons = load_icon_theme(temp_file.path().to_str().unwrap());
-        // Only recording is overridden, others fall back to emoji
-        assert!(icons.idle.contains("🎙"));
-        assert_eq!(icons.recording, "🔴");
-        assert!(icons.transcribing.contains("⏳"));
-    }
-
-    #[test]
-    fn test_context_window_optimization_default_false() {
-        // Default config should have context_window_optimization disabled
-        // (disabled by default due to repetition issues with some models)
-        let config = Config::default();
-        assert!(!config.whisper.context_window_optimization);
-    }
-
-    #[test]
-    fn test_context_window_optimization_can_be_enabled() {
-        let toml_str = r#"
-            [hotkey]
-            key = "SCROLLLOCK"
-
-            [audio]
-            device = "default"
-            sample_rate = 16000
-            max_duration_secs = 60
-
-            [whisper]
-            model = "base.en"
-            language = "en"
-            context_window_optimization = true
-
-            [output]
-            mode = "type"
-        "#;
-
-        let config: Config = toml::from_str(toml_str).unwrap();
-        assert!(config.whisper.context_window_optimization);
-    }
-
-    #[test]
-    fn test_context_window_optimization_defaults_when_omitted() {
-        // When not specified in config, should default to false
-        // (disabled by default due to repetition issues with some models)
-        let toml_str = r#"
-            [hotkey]
-            key = "SCROLLLOCK"
-
-            [audio]
-            device = "default"
-            sample_rate = 16000
-            max_duration_secs = 60
-
-            [whisper]
-            model = "base.en"
-            language = "en"
-
-            [output]
-            mode = "type"
-        "#;
-
-        let config: Config = toml::from_str(toml_str).unwrap();
-        assert!(!config.whisper.context_window_optimization);
-    }
-
-    #[test]
-    fn test_language_config_single() {
-        let toml_str = r#"
-            [whisper]
-            model = "base.en"
-            language = "fr"
-        "#;
-
-        #[derive(Deserialize)]
-        struct TestConfig {
-            whisper: WhisperConfig,
-        }
-
-        let config: TestConfig = toml::from_str(toml_str).unwrap();
-        assert_eq!(
-            config.whisper.language,
-            LanguageConfig::Single("fr".to_string())
-        );
-        assert!(!config.whisper.language.is_auto());
-        assert!(!config.whisper.language.is_multiple());
-        assert_eq!(config.whisper.language.primary(), "fr");
-        assert_eq!(config.whisper.language.as_vec(), vec!["fr"]);
-    }
-
-    #[test]
-    fn test_language_config_auto() {
-        let toml_str = r#"
-            [whisper]
-            model = "large-v3"
-            language = "auto"
-        "#;
-
-        #[derive(Deserialize)]
-        struct TestConfig {
-            whisper: WhisperConfig,
-        }
-
-        let config: TestConfig = toml::from_str(toml_str).unwrap();
-        assert_eq!(
-            config.whisper.language,
-            LanguageConfig::Single("auto".to_string())
-        );
-        assert!(config.whisper.language.is_auto());
-        assert!(!config.whisper.language.is_multiple());
-        assert_eq!(config.whisper.language.primary(), "auto");
-    }
-
-    #[test]
-    fn test_language_config_array() {
-        let toml_str = r#"
-            [whisper]
-            model = "large-v3-turbo"
-            language = ["en", "fr", "de"]
-        "#;
-
-        #[derive(Deserialize)]
-        struct TestConfig {
-            whisper: WhisperConfig,
-        }
-
-        let config: TestConfig = toml::from_str(toml_str).unwrap();
-        assert_eq!(
-            config.whisper.language,
-            LanguageConfig::Multiple(vec!["en".to_string(), "fr".to_string(), "de".to_string()])
-        );
-        assert!(!config.whisper.language.is_auto());
-        assert!(config.whisper.language.is_multiple());
-        assert_eq!(config.whisper.language.primary(), "en");
-        assert_eq!(config.whisper.language.as_vec(), vec!["en", "fr", "de"]);
-    }
-
-    #[test]
-    fn test_language_config_single_element_array() {
-        // A single-element array should not be considered "multiple"
-        let toml_str = r#"
-            [whisper]
-            model = "base.en"
-            language = ["en"]
-        "#;
-
-        #[derive(Deserialize)]
-        struct TestConfig {
-            whisper: WhisperConfig,
-        }
-
-        let config: TestConfig = toml::from_str(toml_str).unwrap();
-        assert!(!config.whisper.language.is_multiple());
-        assert_eq!(config.whisper.language.primary(), "en");
-    }
-
-    #[test]
-    fn test_language_config_default() {
-        // Default should be "en"
-        let config = LanguageConfig::default();
-        assert_eq!(config, LanguageConfig::Single("en".to_string()));
-        assert!(!config.is_auto());
-        assert!(!config.is_multiple());
-        assert_eq!(config.primary(), "en");
-    }
-
-    // =========================================================================
-    // Engine and Mode Tests (v5 config schema)
-    // =========================================================================
-
-    #[test]
-    fn test_parse_engine_whisper() {
-        let toml_str = r#"
-            engine = "whisper"
-
-            [hotkey]
-            key = "SCROLLLOCK"
-
-            [audio]
-            device = "default"
-            sample_rate = 16000
-            max_duration_secs = 60
-
-            [whisper]
-            model = "base.en"
-            language = "en"
-
-            [output]
-            mode = "type"
-        "#;
-
-        let config: Config = toml::from_str(toml_str).unwrap();
-        assert_eq!(config.engine, TranscriptionEngine::Whisper);
-    }
-
-    #[test]
-    fn test_parse_engine_parakeet() {
-        let toml_str = r#"
-            engine = "parakeet"
-
-            [hotkey]
-            key = "SCROLLLOCK"
-
-            [audio]
-            device = "default"
-            sample_rate = 16000
-            max_duration_secs = 60
-
-            [whisper]
-            model = "base.en"
-            language = "en"
-
-            [output]
-            mode = "type"
-
-            [parakeet]
-            model = "parakeet-tdt-0.6b-v3"
-        "#;
-
-        let config: Config = toml::from_str(toml_str).unwrap();
-        assert_eq!(config.engine, TranscriptionEngine::Parakeet);
-        assert!(config.parakeet.is_some());
-        assert_eq!(
-            config.parakeet.as_ref().unwrap().model,
-            "parakeet-tdt-0.6b-v3"
-        );
-    }
-
-    #[test]
-    fn test_engine_defaults_to_whisper() {
-        let toml_str = r#"
-            [hotkey]
-            key = "SCROLLLOCK"
-
-            [audio]
-            device = "default"
-            sample_rate = 16000
-            max_duration_secs = 60
-
-            [whisper]
-            model = "base.en"
-            language = "en"
-
-            [output]
-            mode = "type"
-        "#;
-
-        let config: Config = toml::from_str(toml_str).unwrap();
-        assert_eq!(config.engine, TranscriptionEngine::Whisper);
-    }
-
-    #[test]
-    fn test_output_driver_from_str() {
-        assert_eq!(
-            "wtype".parse::<OutputDriver>().unwrap(),
-            OutputDriver::Wtype
-        );
-        assert_eq!(
-            "dotool".parse::<OutputDriver>().unwrap(),
-            OutputDriver::Dotool
-        );
-        assert_eq!(
-            "ydotool".parse::<OutputDriver>().unwrap(),
-            OutputDriver::Ydotool
-        );
-        assert_eq!(
-            "clipboard".parse::<OutputDriver>().unwrap(),
-            OutputDriver::Clipboard
-        );
-        assert_eq!(
-            "xclip".parse::<OutputDriver>().unwrap(),
-            OutputDriver::Xclip
-        );
-        // Case insensitive
-        assert_eq!(
-            "WTYPE".parse::<OutputDriver>().unwrap(),
-            OutputDriver::Wtype
-        );
-        assert_eq!(
-            "Ydotool".parse::<OutputDriver>().unwrap(),
-            OutputDriver::Ydotool
-        );
-        assert_eq!(
-            "XCLIP".parse::<OutputDriver>().unwrap(),
-            OutputDriver::Xclip
-        );
-        // Invalid
-        assert!("invalid".parse::<OutputDriver>().is_err());
-    }
-
-    #[test]
-    fn test_output_driver_display() {
-        assert_eq!(OutputDriver::Wtype.to_string(), "wtype");
-        assert_eq!(OutputDriver::Dotool.to_string(), "dotool");
-        assert_eq!(OutputDriver::Ydotool.to_string(), "ydotool");
-        assert_eq!(OutputDriver::Clipboard.to_string(), "clipboard");
-        assert_eq!(OutputDriver::Xclip.to_string(), "xclip");
-    }
-
-    #[test]
-    fn test_parse_driver_order_from_toml() {
-        let toml_str = r#"
-            [hotkey]
-            key = "SCROLLLOCK"
-
-            [audio]
-            device = "default"
-            sample_rate = 16000
-            max_duration_secs = 60
-
-            [whisper]
-            model = "base.en"
-            language = "en"
-
-            [output]
-            mode = "type"
-            driver_order = ["ydotool", "wtype", "clipboard"]
-        "#;
-
-        let config: Config = toml::from_str(toml_str).unwrap();
-        let driver_order = config.output.driver_order.unwrap();
-        assert_eq!(driver_order.len(), 3);
-        assert_eq!(driver_order[0], OutputDriver::Ydotool);
-        assert_eq!(driver_order[1], OutputDriver::Wtype);
-        assert_eq!(driver_order[2], OutputDriver::Clipboard);
-    }
-
-    #[test]
-    fn test_parse_whisper_mode_local() {
-        let toml_str = r#"
-            [hotkey]
-            key = "SCROLLLOCK"
-
-            [audio]
-            device = "default"
-            sample_rate = 16000
-            max_duration_secs = 60
-
-            [whisper]
-            mode = "local"
-            model = "base.en"
-            language = "en"
-
-            [output]
-            mode = "type"
-        "#;
-
-        let config: Config = toml::from_str(toml_str).unwrap();
-        assert_eq!(config.whisper.mode, Some(WhisperMode::Local));
-        assert_eq!(config.whisper.effective_mode(), WhisperMode::Local);
-    }
-
-    #[test]
-    fn test_parse_whisper_mode_remote() {
-        let toml_str = r#"
-            [hotkey]
-            key = "SCROLLLOCK"
-
-            [audio]
-            device = "default"
-            sample_rate = 16000
-            max_duration_secs = 60
-
-            [whisper]
-            mode = "remote"
-            model = "base.en"
-            language = "en"
-            remote_endpoint = "http://localhost:8080"
-
-            [output]
-            mode = "type"
-        "#;
-
-        let config: Config = toml::from_str(toml_str).unwrap();
-        assert_eq!(config.whisper.mode, Some(WhisperMode::Remote));
-        assert_eq!(config.whisper.effective_mode(), WhisperMode::Remote);
-    }
-
-    #[test]
-    fn test_whisper_backend_alias_local() {
-        // Test that deprecated 'backend' field still works
-        let toml_str = r#"
-            [hotkey]
-            key = "SCROLLLOCK"
-
-            [audio]
-            device = "default"
-            sample_rate = 16000
-            max_duration_secs = 60
-
-            [whisper]
-            backend = "local"
-            model = "base.en"
-            language = "en"
-
-            [output]
-            mode = "type"
-        "#;
-
-        let config: Config = toml::from_str(toml_str).unwrap();
-        assert_eq!(config.whisper.backend, Some(WhisperMode::Local));
-        assert!(config.whisper.mode.is_none());
-        // effective_mode should return the backend value
-        assert_eq!(config.whisper.effective_mode(), WhisperMode::Local);
-    }
-
-    #[test]
-    fn test_whisper_backend_alias_remote() {
-        // Test that deprecated 'backend' field still works for remote
-        let toml_str = r#"
-            [hotkey]
-            key = "SCROLLLOCK"
-
-            [audio]
-            device = "default"
-            sample_rate = 16000
-            max_duration_secs = 60
-
-            [whisper]
-            backend = "remote"
-            model = "base.en"
-            language = "en"
-            remote_endpoint = "http://localhost:8080"
-
-            [output]
-            mode = "type"
-        "#;
-
-        let config: Config = toml::from_str(toml_str).unwrap();
-        assert_eq!(config.whisper.backend, Some(WhisperMode::Remote));
-        assert!(config.whisper.mode.is_none());
-        // effective_mode should return the backend value
-        assert_eq!(config.whisper.effective_mode(), WhisperMode::Remote);
-    }
-
-    #[test]
-    fn test_whisper_mode_takes_precedence_over_backend() {
-        // When both mode and backend are set, mode should take precedence
-        let toml_str = r#"
-            [hotkey]
-            key = "SCROLLLOCK"
-
-            [audio]
-            device = "default"
-            sample_rate = 16000
-            max_duration_secs = 60
-
-            [whisper]
-            mode = "local"
-            backend = "remote"
-            model = "base.en"
-            language = "en"
-
-            [output]
-            mode = "type"
-        "#;
-
-        let config: Config = toml::from_str(toml_str).unwrap();
-        assert_eq!(config.whisper.mode, Some(WhisperMode::Local));
-        assert_eq!(config.whisper.backend, Some(WhisperMode::Remote));
-        // mode takes precedence
-        assert_eq!(config.whisper.effective_mode(), WhisperMode::Local);
-    }
-
-    #[test]
-    fn test_whisper_effective_mode_defaults_to_local() {
-        // When neither mode nor backend is set, effective_mode defaults to Local
-        let toml_str = r#"
-            [hotkey]
-            key = "SCROLLLOCK"
-
-            [audio]
-            device = "default"
-            sample_rate = 16000
-            max_duration_secs = 60
-
-            [whisper]
-            model = "base.en"
-            language = "en"
-
-            [output]
-            mode = "type"
-        "#;
-
-        let config: Config = toml::from_str(toml_str).unwrap();
-        assert!(config.whisper.mode.is_none());
-        assert!(config.whisper.backend.is_none());
-        assert_eq!(config.whisper.effective_mode(), WhisperMode::Local);
-    }
-
-    // =========================================================================
-    // ParakeetConfig and ParakeetModelType Tests
-    // =========================================================================
-
-    #[test]
-    fn test_parse_parakeet_model_type_tdt() {
-        let toml_str = r#"
-            engine = "parakeet"
-
-            [hotkey]
-            key = "SCROLLLOCK"
-
-            [audio]
-            device = "default"
-            sample_rate = 16000
-            max_duration_secs = 60
-
-            [whisper]
-            model = "base.en"
-            language = "en"
-
-            [output]
-            mode = "type"
-
-            [parakeet]
-            model = "parakeet-tdt-0.6b-v3"
-            model_type = "tdt"
-        "#;
-
-        let config: Config = toml::from_str(toml_str).unwrap();
-        let parakeet = config.parakeet.unwrap();
-        assert_eq!(parakeet.model, "parakeet-tdt-0.6b-v3");
-        assert_eq!(parakeet.model_type, Some(ParakeetModelType::Tdt));
-    }
-
-    #[test]
-    fn test_parse_parakeet_model_type_ctc() {
-        let toml_str = r#"
-            engine = "parakeet"
-
-            [hotkey]
-            key = "SCROLLLOCK"
-
-            [audio]
-            device = "default"
-            sample_rate = 16000
-            max_duration_secs = 60
-
-            [whisper]
-            model = "base.en"
-            language = "en"
-
-            [output]
-            mode = "type"
-
-            [parakeet]
-            model = "parakeet-ctc-0.6b"
-            model_type = "ctc"
-        "#;
-
-        let config: Config = toml::from_str(toml_str).unwrap();
-        let parakeet = config.parakeet.unwrap();
-        assert_eq!(parakeet.model, "parakeet-ctc-0.6b");
-        assert_eq!(parakeet.model_type, Some(ParakeetModelType::Ctc));
-    }
-
-    #[test]
-    fn test_parakeet_model_type_defaults_to_none_for_auto_detection() {
-        let toml_str = r#"
-            engine = "parakeet"
-
-            [hotkey]
-            key = "SCROLLLOCK"
-
-            [audio]
-            device = "default"
-            sample_rate = 16000
-            max_duration_secs = 60
-
-            [whisper]
-            model = "base.en"
-            language = "en"
-
-            [output]
-            mode = "type"
-
-            [parakeet]
-            model = "parakeet-tdt-0.6b-v3"
-        "#;
-
-        let config: Config = toml::from_str(toml_str).unwrap();
-        let parakeet = config.parakeet.unwrap();
-        // model_type should be None (will be auto-detected at runtime)
-        assert!(parakeet.model_type.is_none());
-    }
-
-    #[test]
-    fn test_parakeet_config_default() {
-        let config = ParakeetConfig::default();
-        assert_eq!(config.model, "parakeet-tdt-0.6b-v3");
-        assert!(config.model_type.is_none());
-        assert!(!config.on_demand_loading);
-    }
-
-    #[test]
-    fn test_parakeet_model_type_enum_default() {
-        // ParakeetModelType defaults to Tdt
-        assert_eq!(ParakeetModelType::default(), ParakeetModelType::Tdt);
-    }
-
-    #[test]
-    fn test_whisper_section_is_optional() {
-        // The [whisper] section should be optional for Parakeet users
-        // See: https://github.com/peteonrails/voxtype/issues/137
-        //
-        // We test this by deserializing into a struct that mirrors Config
-        // but only has the fields we want to test (avoiding all required fields)
-        #[derive(Debug, Deserialize)]
-        struct PartialConfig {
-            engine: TranscriptionEngine,
-            #[serde(default)]
-            whisper: WhisperConfig,
-        }
-
-        let toml = r#"
-            engine = "parakeet"
-        "#;
-
-        let config: PartialConfig =
-            toml::from_str(toml).expect("whisper section should be optional");
-        assert_eq!(config.engine, TranscriptionEngine::Parakeet);
-        assert_eq!(config.whisper.model, "base.en"); // Default value
-    }
-
-    #[test]
-    fn test_config_on_demand_loading_whisper() {
-        let config = Config::default();
-        assert_eq!(config.engine, TranscriptionEngine::Whisper);
-        // on_demand_loading method should return whisper's value
-        assert!(!config.on_demand_loading());
-    }
-
-    #[test]
-    fn test_config_model_name_whisper() {
-        let config = Config::default();
-        assert_eq!(config.model_name(), "base.en");
-    }
-
-    // =========================================================================
-    // Profile Tests
-    // =========================================================================
-    #[test]
-    fn test_profiles_default_empty() {
-        let config = Config::default();
-        assert!(config.profiles.is_empty());
-        assert!(config.profile_names().is_empty());
-        assert!(config.get_profile("slack").is_none());
-    }
-
-    #[test]
-    fn test_parse_profiles_from_toml() {
-        let toml_str = r#"
-            [hotkey]
-            key = "SCROLLLOCK"
-
-            [audio]
-            device = "default"
-            sample_rate = 16000
-            max_duration_secs = 60
-
-            [whisper]
-            model = "base.en"
-            language = "en"
-
-            [output]
-            mode = "type"
-
-            [profiles.slack]
-            post_process_command = "cleanup-for-slack.sh"
-
-            [profiles.code]
-            post_process_command = "cleanup-for-code.sh"
-            output_mode = "clipboard"
-        "#;
-
-        let config: Config = toml::from_str(toml_str).unwrap();
-        assert_eq!(config.profiles.len(), 2);
-
-        let slack = config.get_profile("slack").unwrap();
-        assert_eq!(
-            slack.post_process_command,
-            Some("cleanup-for-slack.sh".to_string())
-        );
-        assert!(slack.output_mode.is_none());
-
-        let code = config.get_profile("code").unwrap();
-        assert_eq!(
-            code.post_process_command,
-            Some("cleanup-for-code.sh".to_string())
-        );
-        assert_eq!(code.output_mode, Some(OutputMode::Clipboard));
-    }
-
-    #[test]
-    fn test_parse_profile_with_timeout() {
-        let toml_str = r#"
-            [hotkey]
-            key = "SCROLLLOCK"
-
-            [audio]
-            device = "default"
-            sample_rate = 16000
-            max_duration_secs = 60
-
-            [whisper]
-            model = "base.en"
-            language = "en"
-
-            [output]
-            mode = "type"
-
-            [profiles.slow]
-            post_process_command = "slow-llm-command"
-            post_process_timeout_ms = 60000
-        "#;
-
-        let config: Config = toml::from_str(toml_str).unwrap();
-        let slow = config.get_profile("slow").unwrap();
-        assert_eq!(
-            slow.post_process_command,
-            Some("slow-llm-command".to_string())
-        );
-        assert_eq!(slow.post_process_timeout_ms, Some(60000));
-    }
-
-    #[test]
-    fn test_profile_names() {
-        let toml_str = r#"
-            [hotkey]
-            key = "SCROLLLOCK"
-
-            [audio]
-            device = "default"
-            sample_rate = 16000
-            max_duration_secs = 60
-
-            [whisper]
-            model = "base.en"
-            language = "en"
-
-            [output]
-            mode = "type"
-
-            [profiles.alpha]
-            post_process_command = "alpha-cmd"
-
-            [profiles.beta]
-            post_process_command = "beta-cmd"
-
-            [profiles.gamma]
-            post_process_command = "gamma-cmd"
-        "#;
-
-        let config: Config = toml::from_str(toml_str).unwrap();
-        let names: Vec<&str> = config.profile_names().iter().map(|s| s.as_str()).collect();
-        assert_eq!(names.len(), 3);
-        assert!(names.contains(&"alpha"));
-        assert!(names.contains(&"beta"));
-        assert!(names.contains(&"gamma"));
-    }
-
-    #[test]
-    fn test_profile_without_post_process_command() {
-        // A profile can have only output_mode override without post_process_command
-        let toml_str = r#"
-            [hotkey]
-            key = "SCROLLLOCK"
-
-            [audio]
-            device = "default"
-            sample_rate = 16000
-            max_duration_secs = 60
-
-            [whisper]
-            model = "base.en"
-            language = "en"
-
-            [output]
-            mode = "type"
-
-            [profiles.clipboard_only]
-            output_mode = "clipboard"
-        "#;
-
-        let config: Config = toml::from_str(toml_str).unwrap();
-        let profile = config.get_profile("clipboard_only").unwrap();
-        assert!(profile.post_process_command.is_none());
-        assert_eq!(profile.output_mode, Some(OutputMode::Clipboard));
-    }
-
-    #[test]
-    fn test_config_without_profiles_section() {
-        // Config without [profiles] section should work (backwards compatibility)
-        let toml_str = r#"
-            [hotkey]
-            key = "SCROLLLOCK"
-
-            [audio]
-            device = "default"
-            sample_rate = 16000
-            max_duration_secs = 60
-
-            [whisper]
-            model = "base.en"
-            language = "en"
-
-            [output]
-            mode = "type"
-        "#;
-
-        let config: Config = toml::from_str(toml_str).unwrap();
-        assert!(config.profiles.is_empty());
-    }
-
-    #[test]
-    fn test_parse_driver_order_from_config() {
-        let toml_str = r#"
-            [hotkey]
-            key = "SCROLLLOCK"
-
-            [audio]
-            device = "default"
-            sample_rate = 16000
-            max_duration_secs = 60
-
-            [whisper]
-            model = "base.en"
-            language = "en"
-
-            [output]
-            mode = "type"
-            driver_order = ["ydotool", "wtype", "clipboard"]
-        "#;
-
-        let config: Config = toml::from_str(toml_str).unwrap();
-        let driver_order = config.output.driver_order.unwrap();
-        assert_eq!(driver_order.len(), 3);
-        assert_eq!(driver_order[0], OutputDriver::Ydotool);
-        assert_eq!(driver_order[1], OutputDriver::Wtype);
-        assert_eq!(driver_order[2], OutputDriver::Clipboard);
-    }
-
-    #[test]
-    fn test_driver_order_not_set_by_default() {
-        let config = Config::default();
-        assert!(config.output.driver_order.is_none());
-    }
-
-    #[test]
-    fn test_parse_config_without_driver_order() {
-        // Ensure backwards compatibility - config without driver_order should work
-        let toml_str = r#"
-            [hotkey]
-            key = "SCROLLLOCK"
-
-            [audio]
-            device = "default"
-            sample_rate = 16000
-            max_duration_secs = 60
-
-            [whisper]
-            model = "base.en"
-            language = "en"
-
-            [output]
-            mode = "type"
-        "#;
-
-        let config: Config = toml::from_str(toml_str).unwrap();
-        assert!(config.output.driver_order.is_none());
-    }
-
-    #[test]
-    fn test_parse_single_driver_order() {
-        let toml_str = r#"
-            [hotkey]
-            key = "SCROLLLOCK"
-
-            [audio]
-            device = "default"
-            sample_rate = 16000
-            max_duration_secs = 60
-
-            [whisper]
-            model = "base.en"
-            language = "en"
-
-            [output]
-            mode = "type"
-            driver_order = ["ydotool"]
-        "#;
-
-        let config: Config = toml::from_str(toml_str).unwrap();
-        let driver_order = config.output.driver_order.unwrap();
-        assert_eq!(driver_order.len(), 1);
-        assert_eq!(driver_order[0], OutputDriver::Ydotool);
-    }
-
-    // =========================================================================
-    // Meeting Config Tests
-    // =========================================================================
-
-    #[test]
-    fn test_meeting_config_default() {
-        let config = MeetingConfig::default();
-        assert!(!config.enabled);
-        assert_eq!(config.chunk_duration_secs, 30);
-        assert_eq!(config.storage_path, "auto");
-        assert!(!config.retain_audio);
-        assert_eq!(config.max_duration_mins, 180);
-    }
-
-    #[test]
-    fn test_meeting_audio_config_default() {
-        let config = MeetingAudioConfig::default();
-        assert_eq!(config.mic_device, "default");
-        assert_eq!(config.loopback_device, "auto");
-        assert_eq!(config.vad_threshold, 0.01);
-    }
-
-    #[test]
-    fn test_meeting_diarization_config_default() {
-        let config = MeetingDiarizationConfig::default();
-        assert!(config.enabled);
-        assert_eq!(config.backend, "simple");
-        assert_eq!(config.max_speakers, 10);
-    }
-
-    #[test]
-    fn test_meeting_summary_config_default() {
-        let config = MeetingSummaryConfig::default();
-        assert_eq!(config.backend, "disabled");
-        assert_eq!(config.ollama_url, "http://localhost:11434");
-        assert_eq!(config.ollama_model, "llama3.2");
-        assert!(config.remote_endpoint.is_none());
-        assert!(config.remote_api_key.is_none());
-        assert_eq!(config.timeout_secs, 120);
-    }
-
-    #[test]
-    fn test_meeting_config_in_default_config() {
-        let config = Config::default();
-        assert!(!config.meeting.enabled);
-        assert_eq!(config.meeting.chunk_duration_secs, 30);
-        assert_eq!(config.meeting.max_duration_mins, 180);
-    }
-
-    #[test]
-    fn test_parse_meeting_config_from_toml() {
-        let toml_str = r#"
-            [hotkey]
-            key = "SCROLLLOCK"
-
-            [audio]
-            device = "default"
-            sample_rate = 16000
-            max_duration_secs = 60
-
-            [whisper]
-            model = "base.en"
-            language = "en"
-
-            [output]
-            mode = "type"
-
-            [meeting]
-            enabled = true
-            chunk_duration_secs = 45
-            storage_path = "/tmp/meetings"
-            retain_audio = true
-            max_duration_mins = 60
-        "#;
-
-        let config: Config = toml::from_str(toml_str).unwrap();
-        assert!(config.meeting.enabled);
-        assert_eq!(config.meeting.chunk_duration_secs, 45);
-        assert_eq!(config.meeting.storage_path, "/tmp/meetings");
-        assert!(config.meeting.retain_audio);
-        assert_eq!(config.meeting.max_duration_mins, 60);
-    }
-
-    #[test]
-    fn test_parse_meeting_config_with_nested_sections() {
-        let toml_str = r#"
-            [hotkey]
-            key = "SCROLLLOCK"
-
-            [audio]
-            device = "default"
-            sample_rate = 16000
-            max_duration_secs = 60
-
-            [whisper]
-            model = "base.en"
-            language = "en"
-
-            [output]
-            mode = "type"
-
-            [meeting]
-            enabled = true
-
-            [meeting.audio]
-            mic_device = "hw:1"
-            loopback_device = "disabled"
-            vad_threshold = 0.001
-
-            [meeting.diarization]
-            enabled = false
-            backend = "ml"
-            max_speakers = 5
-
-            [meeting.summary]
-            backend = "local"
-            ollama_model = "mistral"
-            timeout_secs = 60
-        "#;
-
-        let config: Config = toml::from_str(toml_str).unwrap();
-        assert_eq!(config.meeting.audio.mic_device, "hw:1");
-        assert_eq!(config.meeting.audio.loopback_device, "disabled");
-        assert_eq!(config.meeting.audio.vad_threshold, 0.001);
-        assert!(!config.meeting.diarization.enabled);
-        assert_eq!(config.meeting.diarization.backend, "ml");
-        assert_eq!(config.meeting.diarization.max_speakers, 5);
-        assert_eq!(config.meeting.summary.backend, "local");
-        assert_eq!(config.meeting.summary.ollama_model, "mistral");
-        assert_eq!(config.meeting.summary.timeout_secs, 60);
-    }
-
-    #[test]
-    fn test_meeting_config_backward_compatible_omitted() {
-        // Config without [meeting] section should parse fine with defaults
-        let toml_str = r#"
-            [hotkey]
-            key = "SCROLLLOCK"
-
-            [audio]
-            device = "default"
-            sample_rate = 16000
-            max_duration_secs = 60
-
-            [whisper]
-            model = "base.en"
-            language = "en"
-
-            [output]
-            mode = "type"
-        "#;
-
-        let config: Config = toml::from_str(toml_str).unwrap();
-        assert!(!config.meeting.enabled);
-        assert_eq!(config.meeting.chunk_duration_secs, 30);
-        assert_eq!(config.meeting.storage_path, "auto");
-        assert_eq!(config.meeting.diarization.backend, "simple");
-        assert_eq!(config.meeting.summary.backend, "disabled");
-    }
-
-    // =========================================================================
-    // Clipboard Restore Tests
-    // =========================================================================
-
-    #[test]
-    fn test_restore_clipboard_defaults() {
-        let config = Config::default();
-        assert!(!config.output.restore_clipboard);
-        assert_eq!(config.output.restore_clipboard_delay_ms, 200);
-    }
-
-    #[test]
-    fn test_restore_clipboard_deserialization() {
-        let toml_str = r#"
-            [hotkey]
-            key = "SCROLLLOCK"
-
-            [audio]
-            device = "default"
-            sample_rate = 16000
-            max_duration_secs = 30
-
-            [whisper]
-            model = "base.en"
-
-            [output]
-            mode = "paste"
-            restore_clipboard = true
-            restore_clipboard_delay_ms = 500
-        "#;
-
-        let config: Config = toml::from_str(toml_str).unwrap();
-        assert!(config.output.restore_clipboard);
-        assert_eq!(config.output.restore_clipboard_delay_ms, 500);
-    }
-
-    #[test]
-    fn test_restore_clipboard_missing_uses_defaults() {
-        let toml_str = r#"
-            [hotkey]
-            key = "SCROLLLOCK"
-
-            [audio]
-            device = "default"
-            sample_rate = 16000
-            max_duration_secs = 30
-
-            [whisper]
-            model = "base.en"
-
-            [output]
-            mode = "paste"
-        "#;
-
-        let config: Config = toml::from_str(toml_str).unwrap();
-        assert!(!config.output.restore_clipboard);
-        assert_eq!(config.output.restore_clipboard_delay_ms, 200);
-    }
-
-    #[test]
-    fn test_parse_profile_modifiers() {
-        let toml_str = r#"
-            [hotkey]
-            key = "SCROLLLOCK"
-
-            [hotkey.profile_modifiers]
-            LEFTSHIFT = "translate"
-            RIGHTALT = "formal"
-
-            [audio]
-            device = "default"
-            sample_rate = 16000
-            max_duration_secs = 60
-
-            [whisper]
-            model = "base.en"
-            language = "en"
-
-            [output]
-            mode = "type"
-
-            [profiles.translate]
-            post_process_command = "translate.sh"
-
-            [profiles.formal]
-            post_process_command = "formal.sh"
-            post_process_timeout_ms = 15000
-        "#;
-
-        let config: Config = toml::from_str(toml_str).unwrap();
-        assert_eq!(config.hotkey.profile_modifiers.len(), 2);
-        assert_eq!(
-            config.hotkey.profile_modifiers.get("LEFTSHIFT").unwrap(),
-            "translate"
-        );
-        assert_eq!(
-            config.hotkey.profile_modifiers.get("RIGHTALT").unwrap(),
-            "formal"
-        );
-        assert!(config.get_profile("translate").is_some());
-        assert!(config.get_profile("formal").is_some());
-        assert_eq!(
-            config
-                .get_profile("translate")
-                .unwrap()
-                .post_process_command
-                .as_deref(),
-            Some("translate.sh")
-        );
-    }
-
-    #[test]
-    fn test_profile_modifiers_default_empty() {
-        let toml_str = r#"
-            [hotkey]
-            key = "SCROLLLOCK"
-
-            [audio]
-            device = "default"
-            sample_rate = 16000
-            max_duration_secs = 60
-
-            [whisper]
-            model = "base.en"
-            language = "en"
-
-            [output]
-            mode = "type"
-        "#;
-
-        let config: Config = toml::from_str(toml_str).unwrap();
-        assert!(config.hotkey.profile_modifiers.is_empty());
-    }
-
-    #[test]
-    fn test_system_path_constant() {
-        assert_eq!(
-            Config::system_path(),
-            PathBuf::from("/etc/voxtype/config.toml")
-        );
-        assert_eq!(Config::SYSTEM_PATH, "/etc/voxtype/config.toml");
-    }
-
-    #[test]
     fn test_load_config_explicit_path() {
         // Explicit --config should always be used regardless of fallback.
         let dir = tempfile::tempdir().unwrap();
@@ -2323,199 +801,13 @@ mod tests {
     }
 
     #[test]
-    fn test_default_language_to_layout_common_cases() {
-        let map = default_language_to_layout();
-        // English maps to "us", the XKB convention.
-        assert_eq!(map.get("en"), Some(&"us".to_string()));
-        // Russian, German, French, Spanish are direct passthroughs.
-        assert_eq!(map.get("ru"), Some(&"ru".to_string()));
-        assert_eq!(map.get("de"), Some(&"de".to_string()));
-        assert_eq!(map.get("fr"), Some(&"fr".to_string()));
-        assert_eq!(map.get("es"), Some(&"es".to_string()));
-        // Greek uses "gr", not "el".
-        assert_eq!(map.get("el"), Some(&"gr".to_string()));
-        // Japanese / Korean map to common XKB names.
-        assert_eq!(map.get("ja"), Some(&"jp".to_string()));
-        assert_eq!(map.get("ko"), Some(&"kr".to_string()));
-    }
-
-    #[test]
-    fn test_output_config_default_includes_language_layout_map() {
-        let cfg = Config::default();
-        assert!(!cfg.output.language_to_layout.is_empty());
+    fn test_system_path_constant() {
         assert_eq!(
-            cfg.output.language_to_layout.get("en"),
-            Some(&"us".to_string())
+            Config::system_path(),
+            PathBuf::from("/etc/voxtype/config.toml")
         );
-        assert!(cfg.output.language_to_variant.is_empty());
-        // New eitype layout fields are unset by default; the layout is
-        // inferred from the detected language only when both fields are
-        // empty (see daemon::handle_transcription_result).
-        assert!(cfg.output.eitype_xkb_layout.is_none());
-        assert!(cfg.output.eitype_xkb_variant.is_none());
+        assert_eq!(Config::SYSTEM_PATH, "/etc/voxtype/config.toml");
     }
-
-    #[test]
-    fn test_parse_eitype_layout_from_toml() {
-        let toml_str = r#"
-            [hotkey]
-            key = "PAUSE"
-
-            [audio]
-            device = "default"
-            sample_rate = 16000
-            max_duration_secs = 60
-
-            [whisper]
-            model = "base.en"
-            language = "en"
-
-            [output]
-            mode = "type"
-            eitype_xkb_layout = "ru"
-            eitype_xkb_variant = "phonetic"
-        "#;
-        let config: Config = toml::from_str(toml_str).unwrap();
-        assert_eq!(config.output.eitype_xkb_layout, Some("ru".to_string()));
-        assert_eq!(
-            config.output.eitype_xkb_variant,
-            Some("phonetic".to_string())
-        );
-    }
-
-    #[test]
-    fn test_parse_language_to_layout_override() {
-        // User can override individual mappings (e.g. Brazilian Portuguese
-        // typically needs the `br` layout, not `pt`). Providing the field
-        // replaces the built-in defaults; users are expected to copy
-        // entries they want to keep (documented in CONFIGURATION.md).
-        let toml_str = r#"
-            [hotkey]
-            key = "PAUSE"
-
-            [audio]
-            device = "default"
-            sample_rate = 16000
-            max_duration_secs = 60
-
-            [whisper]
-            model = "base.en"
-            language = "en"
-
-            [output]
-            mode = "type"
-
-            [output.language_to_layout]
-            pt = "br"
-            en = "dvorak"
-        "#;
-        let config: Config = toml::from_str(toml_str).unwrap();
-        assert_eq!(
-            config.output.language_to_layout.get("pt"),
-            Some(&"br".to_string())
-        );
-        assert_eq!(
-            config.output.language_to_layout.get("en"),
-            Some(&"dvorak".to_string())
-        );
-    }
-
-    #[test]
-    fn test_parse_language_to_variant() {
-        let toml_str = r#"
-            [hotkey]
-            key = "PAUSE"
-
-            [audio]
-            device = "default"
-            sample_rate = 16000
-            max_duration_secs = 60
-
-            [whisper]
-            model = "base.en"
-            language = ["en", "ru"]
-
-            [output]
-            mode = "type"
-
-            [output.language_to_layout]
-            en = "us"
-            ru = "ru"
-
-            [output.language_to_variant]
-            ru = "phonetic"
-        "#;
-        let config: Config = toml::from_str(toml_str).unwrap();
-        assert_eq!(
-            config.output.language_to_variant.get("ru"),
-            Some(&"phonetic".to_string())
-        );
-        assert!(!config.output.language_to_variant.contains_key("en"));
-    }
-
-    #[test]
-    fn test_apply_language_xkb_hint_applies_layout_and_variant() {
-        let mut output = Config::default().output;
-        output
-            .language_to_variant
-            .insert("ru".to_string(), "phonetic".to_string());
-
-        let applied = output.apply_language_xkb_hint("ru");
-
-        assert_eq!(applied.layout, Some("ru".to_string()));
-        assert_eq!(applied.variant, Some("phonetic".to_string()));
-        assert!(applied.eitype_layout_applied);
-        assert!(applied.dotool_layout_applied);
-        assert!(applied.eitype_variant_applied);
-        assert!(applied.dotool_variant_applied);
-        assert_eq!(output.eitype_xkb_layout, Some("ru".to_string()));
-        assert_eq!(output.dotool_xkb_layout, Some("ru".to_string()));
-        assert_eq!(output.eitype_xkb_variant, Some("phonetic".to_string()));
-        assert_eq!(output.dotool_xkb_variant, Some("phonetic".to_string()));
-    }
-
-    #[test]
-    fn test_apply_language_xkb_hint_does_not_leak_variant_between_languages() {
-        let mut output = Config::default().output;
-        output
-            .language_to_variant
-            .insert("ru".to_string(), "phonetic".to_string());
-
-        let applied = output.apply_language_xkb_hint("en");
-
-        assert_eq!(applied.layout, Some("us".to_string()));
-        assert_eq!(applied.variant, None);
-        assert_eq!(output.eitype_xkb_layout, Some("us".to_string()));
-        assert_eq!(output.dotool_xkb_layout, Some("us".to_string()));
-        assert_eq!(output.eitype_xkb_variant, None);
-        assert_eq!(output.dotool_xkb_variant, None);
-    }
-
-    #[test]
-    fn test_apply_language_xkb_hint_preserves_explicit_variant() {
-        let mut output = Config::default().output;
-        output.eitype_xkb_variant = Some("explicit-eitype".to_string());
-        output.dotool_xkb_variant = Some("explicit-dotool".to_string());
-        output
-            .language_to_variant
-            .insert("ru".to_string(), "phonetic".to_string());
-
-        let applied = output.apply_language_xkb_hint("ru");
-
-        assert_eq!(applied.variant, Some("phonetic".to_string()));
-        assert!(!applied.eitype_variant_applied);
-        assert!(!applied.dotool_variant_applied);
-        assert_eq!(
-            output.eitype_xkb_variant,
-            Some("explicit-eitype".to_string())
-        );
-        assert_eq!(
-            output.dotool_xkb_variant,
-            Some("explicit-dotool".to_string())
-        );
-    }
-
-    // ----- Partial-config layering (regression tests for #421) -----
 
     #[test]
     fn empty_toml_yields_default_config() {
